+++ Tests/helpers/New-TestCertificate.ps1
@@ -1,36 +1,22 @@
 function New-TestCertificate
 {
     [CmdletBinding()]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSAvoidUsingConvertToSecureStringWithPlainText", "")]
     param ()
 
-    # Create self signed certificate
-    $certificatePath = "Cert:\LocalMachine\My"
-    $certificate = Get-ChildItem -Path $certificatePath | Where-Object { ($_.Subject -eq "CN=testcert") } | Select-Object -First 1
+    # Create a code signing cert
+    $myCert = New-SelfSignedCertificate -Type 'CodeSigningCert' -DnsName 'GCModuleTestOnly' -HashAlgorithm 'SHA256'
 
-    if ($null -eq $certificate)
-    {
-        $newSelfsignedCertificateExParams = @{
-            Subject            = "CN=testcert"
-            EKU                = 'Code Signing'
-            KeyUsage           = 'KeyEncipherment, DataEncipherment, DigitalSignature'
-            SAN                = $env:ComputerName
-            FriendlyName       = 'DSC Credential Encryption certificate'
-            Exportable         = $true
-            StoreLocation      = 'LocalMachine'
-            KeyLength          = 2048
-            ProviderName       = 'Microsoft Enhanced Cryptographic Provider v1.0'
-            AlgorithmName      = 'RSA'
-            SignatureAlgorithm = 'SHA256'
-        }
+    # Export the certificates
 
-        $null = New-SelfsignedCertificateEx @newSelfsignedCertificateExParams
+    $myPwd = ConvertTo-SecureString -String 'Password1234' -Force -AsPlainText
+    $myCert | Export-PfxCertificate -FilePath 'C:\demo\GCPrivateKey.pfx' -Password $myPwd
+    $myCert | Export-Certificate -FilePath 'C:\demo\GCPublicKey.cer' -Force
 
-    }
+    # Import the certificate
+    $null = Import-PfxCertificate -FilePath 'C:\demo\GCPrivateKey.pfx' -Password $myPwd -CertStoreLocation 'Cert:\LocalMachine\My'
 
-    $command = @'
-$Cert = Get-ChildItem -Path Cert:\LocalMachine\My | Where-Object { ($_.Subject -eq 'CN=testcert') } | Select-Object -First 1
-Export-Certificate -FilePath "$TestDrive/exported.cer" -Cert $Cert
-Import-Certificate -FilePath "$TestDrive/exported.cer" -CertStoreLocation Cert:\LocalMachine\Root
-'@
-    powershell.exe -NoProfile -NonInteractive -Command $command
+    # Sign the package
+    $certToSignThePackage = Get-ChildItem -Path  'Cert:\LocalMachine\My' | Where-Object { ($_.Subject -eq 'CN=GCModuleTestOnly') }
+    return $certToSignThePackage
 }